+++ braintree/customer.py
@@ -159,6 +159,13 @@ class Customer(Resource):
             {"risk_data": ["customer_browser", "customer_ip"]},
             {"credit_card": CreditCard.create_signature()},
             {"custom_fields": ["__any_key__"]},
+            {"three_d_secure_pass_thru": [
+                "cavv",
+                "ds_transaction_id",
+                "eci_flag",
+                "three_d_secure_version",
+                "xid",
+                ]},
             {"options": [{"paypal": [
                 "payee_email",
                 "order_id",
@@ -174,6 +181,13 @@ class Customer(Resource):
         return [
             "company", "email", "fax", "first_name", "id", "last_name", "phone", "website", "device_data", "device_session_id", "fraud_merchant_id", "payment_method_nonce", "default_payment_method_token",
             {"credit_card": CreditCard.signature("update_via_customer")},
+            {"three_d_secure_pass_thru": [
+                "cavv",
+                "ds_transaction_id",
+                "eci_flag",
+                "three_d_secure_version",
+                "xid",
+                ]},
             {"custom_fields": ["__any_key__"]},
             {"options": [{"paypal": [
                 "payee_email",